+++ src/components/HomeBody/HomeBody.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import LeftBody from './LeftBody.jsx';
-import CenterBody from './CenterBody.jsx';
-import RightBody from './RightBody.jsx';
+import LeftBody from '../LeftBody/LeftBody.jsx';
+import CenterBody from '../CenterBody/CenterBody.jsx';
+import RightBody from '../RightBody/RightBody.jsx';
 
 const HomeBody = ({openModal}) => {
   return (
