+++ src/components/LeftBody/LeftBody.style.js
