+++ src/pages/Home/Home.jsx
@@ -1,9 +1,9 @@
 import audio from '@/assets/audio/homeBGM.mp3';
-import { HomeLayoutStyle } from './HomeLayout.style.js';
-import HomeBody from './HomeBody.jsx';
+import { HomeLayoutStyle } from '../../components/HomeLayout/HomeLayout.style.js';
+import HomeBody from '../../components/HomeBody/HomeBody.jsx';
 import Modal from '../../components/Modal/Modal.jsx';
 import { useState } from 'react';
-import ModalContent from './ModalContent.jsx';
+import ModalContent from '../../components/ModalContent/ModalContent.jsx';
 import { WarningScreen } from '../../components/WarningScreen/WarningScreen.jsx';
 export const Home = () => {
   const [isOpen, setIsOpen] = useState(false);
